% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{define_variables}
\alias{define_variables}
\title{Define the imported variables to be used for downstream analysis.}
\usage{
define_variables(
  data,
  input_date,
  input_temp,
  input_rhum,
  input_wood,
  input_ohm,
  output_date = gen_date,
  output_temp = gen_temp,
  output_rhum = gen_rhum,
  output_wood = gen_wood,
  output_ohm = gen_ohm
)
}
\arguments{
\item{data}{The data frame to define variables from.}

\item{input_date}{Optional. Input column containing in the data frame ###}

\item{input_temp}{Optional. Input column containing temperature in the data frame}

\item{input_rhum}{Optional. Input column containing relative humidity in the data frame}

\item{input_wood}{Optional. Input column containing wood moisture in the data frame}

\item{input_ohm}{Optional. Input column containing ohm in the data frame}
}
\value{
A data frame with added columns defined for downstream analysis.
}
\description{
Define the imported variables to be used for downstream analysis.
}
\examples{
\dontrun{
describe(dataset)
describe(dataset, col1, col2)
}



}
