% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{define_variables_ohm}
\alias{define_variables_ohm}
\title{Define Ohm Value Variable}
\usage{
define_variables_ohm(data, input_ohm, output_ohm = gen_ohm)
}
\arguments{
\item{data}{The input data frame.}

\item{input_ohm}{The name or expression of the Ohm value column.}

\item{output_ohm}{The name of the new Ohm value variable to be created.}
}
\value{
A modified data frame with the new Ohm value variable added.
}
\description{
This function takes a data frame and defines an Ohm value variable based on the provided input
Ohm value column, converting it to a double type if necessary.
}
\examples{
# Example usage:
data_with_ohm <- define_variables_ohm(
  data = my_data, input_ohm = OhmValue,
  output_ohm = gen_ohm
)

}
