% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{define_variables_rhum}
\alias{define_variables_rhum}
\title{Define Relative Humidity Variable}
\usage{
define_variables_rhum(data, input_rhum, output_rhum = gen_rhum)
}
\arguments{
\item{data}{The input data frame.}

\item{input_rhum}{The name or expression of the relative humidity column.}

\item{output_rhum}{The name of the new relative humidity variable to be created.}
}
\value{
A modified data frame with the new relative humidity variable added.
}
\description{
This function takes a data frame and defines a relative humidity variable based on the provided input
relative humidity column, converting it to a double type if necessary.
}
\examples{
# Example usage:
data_with_rhum <- define_variables_rhum(
  data = my_data, input_rhum = RelativeHumidity,
  output_rhum = gen_rhum
)

}
