% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{define_variables_sensorID}
\alias{define_variables_sensorID}
\title{#' Define Date Variable
#'
#' This function takes a data frame and defines a date variable based on the provided input
#' date column, using the specified date format.
#'
#' @param data The input data frame.
#' @param input_date The name or expression of the date column.
#' @param date_format The format of the input date column ("ymd" for Year-Month-Day or "dmy" for Day-Month-Year).
#' @param output_date The name of the new date variable to be created.
#'
#' @import dplyr
#' @importFrom lubridate ymd dmy
#'
#' @return A modified data frame with the new date variable added.
#'
#' @examples
#' # Example usage with "ymd" format:
#' data_with_date <- define_variables_date(
#'   data = my_data, input_date = DateColumn,
#'   date_format = "ymd", output_date = NewDate
#' )
#'
#' # Example usage with "dmy" format:
#' data_with_date <- define_variables_date(
#'   data = my_data, input_date = DateColumn,
#'   date_format = "dmy", output_date = NewDate
#' )
#'
#' @export
#'
define_variables_date <- function(data,
input_date,
output_date = "gen_date") {
if (!missing(input_date) && !is.null(input_date)) {
data \%>\%
mutate(!!sym(output_date) := parsedate::parse_date(.data[\link{input_date}]))
} else {
stop("No date column to define is specified")
}
}
Define Sensor ID Variable}
\usage{
define_variables_sensorID(
  data,
  input_sensor,
  input_port,
  output_sensorID = gen_sensorID
)
}
\arguments{
\item{data}{The input data frame.}

\item{input_sensor}{The name or expression of the sensor column.}

\item{input_port}{The name or expression of the port column (optional).}

\item{output_sensorID}{The name of the new sensor ID variable to be created.}
}
\value{
A modified data frame with the new sensor ID variable added.
}
\description{
This function takes a data frame and defines a sensor ID variable based on the provided input
sensor and port columns. It can create a sensor ID by concatenating the sensor and port values
with an underscore or use the sensor column directly.
}
\examples{
# Example usage with sensor and port columns:
data_with_sensorID <- define_variables_sensorID(
  data = my_data, input_sensor = Sensor,
  input_port = Port, output_sensorID = SensorID
)

# Example usage with only a sensor column:
data_with_sensorID <- define_variables_sensorID(
  data = my_data, input_sensor = Sensor,
  output_sensorID = gen_SensorID
)

}
