% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{define_variables_temp}
\alias{define_variables_temp}
\title{Define Temperature Variable}
\usage{
define_variables_temp(data, input_temp, output_temp = gen_temp)
}
\arguments{
\item{data}{The input data frame.}

\item{input_temp}{The name or expression of the temperature column.}

\item{output_temp}{The name of the new temperature variable to be created.}
}
\value{
A modified data frame with the new temperature variable added.
}
\description{
This function takes a data frame and defines a temperature variable based on the provided input
temperature column, converting it to a double type if necessary.
}
\examples{
# Example usage:
data_with_temperature <- define_variables_temp(
  data = my_data, input_temp = Temperature,
  output_temp = gen_temp
)

}
