% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_variables.R
\name{define_variables_wood}
\alias{define_variables_wood}
\title{Define Wood Moisture Variable}
\usage{
define_variables_wood(data, input_wood, output_wood = gen_wood)
}
\arguments{
\item{data}{The input data frame.}

\item{input_wood}{The name or expression of the wood moisture column.}

\item{output_wood}{The name of the new wood moisture variable to be created.}
}
\value{
A modified data frame with the new wood moisture variable added.
}
\description{
This function takes a data frame and defines a wood moisture variable based on the provided input
wood moisture column, converting it to a double type if necessary.
}
\examples{
# Example usage:
data_with_wood <- define_variables_wood(
  data = my_data, input_wood = WoodMoisture,
  output_wood = gen_wood
)

}
