% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_complete_date.R
\name{make_complete_date}
\alias{make_complete_date}
\title{Make Date Variable Complete with Regular Intervals}
\usage{
make_complete_date(
  data,
  input_date = "gen_date",
  input_site_id = NULL,
  input_sensor_id = NULL,
  input_sensor_ports = NULL,
  timeframe = "hour"
)
}
\arguments{
\item{data}{The input data frame containing the date variable to be completed.}

\item{input_date}{The name or expression of the date variable to complete.}

\item{input_site_id}{The name or expression of the site ID variable.}

\item{input_sensor_id}{The name or expression of the sensor ID variable.}

\item{timeframe}{A character vector specifying the timeframe for regular intervals
(e.g., "hour", "day", "week", "month").}
}
\value{
A modified data frame with the date variable completed with regular intervals.
}
\description{
This function takes a data frame and completes a date variable to fill in missing time gaps
with regular intervals specified by the 'timeframe' argument. It also performs grouping by site
and sensor before the completion and ungroups the dataset prior to return.
}
\examples{
# Example usage:
completed_data <- make_complete_date(
  data = my_data, input_date = Date,
  input_site_id = SiteID, input_sensor_id = SensorID,
  timeframe = "hour"
)

}
