% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mycoindex_mold.R
\name{make_mycoindex_mold}
\alias{make_mycoindex_mold}
\title{Create Mold Index Variable}
\usage{
make_mycoindex_mold(
  data = data,
  input_mold = gen_rhum,
  output_index = MIx_mold,
  mold_low = 75,
  mold_mid = 85,
  mold_high = 95
)
}
\arguments{
\item{data}{The input data frame containing the relative humidity measurements.}

\item{input_mold}{The name or expression of the relative humidity measurement variable.}

\item{output_index}{The name of the new mold index variable to be created.}

\item{mold_low}{The threshold for the lower mold level (e.g., 75).}

\item{mold_mid}{The threshold for the mid-range mold level (e.g., 85).}

\item{mold_high}{The threshold for the high mold level (e.g., 95).}
}
\value{
A modified data frame with the new mold index variable added.
}
\description{
This function takes a data frame and creates a mold index variable based on the provided
relative humidity measurements, using user-defined thresholds for different mold levels.
}
\examples{
# Example usage:
data_with_mold_index <- make_mycoindex_mold(
  data = my_data, input_mold = MoldMeasurement,
  output_index = MoldIndex, mold_low = 75,
  mold_mid = 85, mold_high = 95
)

}
