% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mycoindex_temp.R
\name{make_mycoindex_temp}
\alias{make_mycoindex_temp}
\title{Create Temperature Index Variable}
\usage{
make_mycoindex_temp(
  data = data,
  input_temp = gen_temp,
  output_index = MIx_temp,
  temp_low = 4,
  temp_mid = 8,
  temp_high = 14,
  temp_max = 35
)
}
\arguments{
\item{data}{The input data frame containing the temperature measurements.}

\item{input_temp}{The name or expression of the temperature measurement variable.}

\item{output_index}{The name of the new temperature index variable to be created.}

\item{temp_low}{The threshold for the low temperature level (e.g., 4).}

\item{temp_mid}{The threshold for the mid-range temperature level (e.g., 8).}

\item{temp_high}{The threshold for the high temperature level (e.g., 14).}

\item{temp_max}{The maximum temperature threshold (e.g., 35).}
}
\value{
A modified data frame with the new temperature index variable added.
}
\description{
This function takes a data frame and creates a temperature index variable based on the provided
temperature measurements, using user-defined thresholds for different temperature levels.
}
\examples{
# Example usage:
data_with_temp_index <- make_mycoindex_temp(
  data = my_data, input_temp = Temperature,
  output_index = TempIndex, temp_low = 4,
  temp_mid = 8, temp_high = 14, temp_max = 35
)

}
