% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mycoindex_wood.R
\name{make_mycoindex_wood}
\alias{make_mycoindex_wood}
\title{Create Wood Moisture Index Variable}
\usage{
make_mycoindex_wood(
  data = data,
  input_wood = gen_wood,
  output_index = MIx_wood,
  wood_low = 20,
  wood_mid = 25,
  wood_high = 30,
  wood_max = 100
)
}
\arguments{
\item{data}{The input data frame containing the wood moisture measurements.}

\item{input_wood}{The name or expression of the wood moisture measurement variable.}

\item{output_index}{The name of the new wood moisture index variable to be created.}

\item{wood_low}{The threshold for the low wood moisture level (e.g., 20).}

\item{wood_mid}{The threshold for the mid-range wood moisture level (e.g., 25).}

\item{wood_high}{The threshold for the high wood moisture level (e.g., 30).}

\item{wood_max}{The maximum wood moisture threshold (e.g., 100).}
}
\value{
A modified data frame with the new wood moisture index variable added.
}
\description{
This function takes a data frame and creates a wood moisture index variable based on the provided
wood moisture measurements, using user-defined thresholds for different wood moisture levels.
}
\examples{
# Example usage:
data_with_wood_index <- make_mycoindex_wood(
  data = my_data, input_wood = WoodMoisture,
  output_index = WoodMoistureIndex, wood_low = 20,
  wood_mid = 25, wood_high = 30, wood_max = 100
)

}
