% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rolling.R
\name{make_rolling_mix_wood}
\alias{make_rolling_mix_wood}
\title{Rolling mean for MIx_wood (time-aware)}
\usage{
make_rolling_mix_wood(
  data,
  input = MIx_wood,
  index = gen_datetime,
  id = NULL,
  output_name,
  roll_interval = 24,
  align = "right",
  na_rm = TRUE,
  complete = TRUE
)
}
\description{
Rolling mean for MIx_wood (time-aware)
}
